Rework face constructor to accept SimpleColor-like
Instead of restricting the colour component to be either a SimpleColor
or one of the three types that a SimpleColor constructor is defined
for (RGBTuple, Symbol, and UInt32), we accept anything and then try to
convert to a SimpleColor.

This is almost functionally equivalent, but it allows other
packages (e.g. Colors.jl) to define their own convert(SimpleColor, x)
methods, which will then "just work" with Faces.