Make loadfaces! singular for addface! consistency
Yes, there the plural is more apt with the Dict method, but that's only
really expected to be used when loading a Face.toml. I.e. most of the
time `loadface!` will be applied to a single face, and so the singular
for is more appropriate.

Maybe the Dict method should be renamed entirely in the future?