Add documentation task to CI
Before now, documentation build errors would only be caught when testing
manually or PRing a new version of StyledStrings.jl to the main Julia
repo.

To address this, a new CI task is introduced, and a few small changes
made to make automated docs builds possible, namely:
- Tweaking the documentation workflow
- Introducing some docs build scripts under docs/
- Fixing a dodgy jldoctest

We also add a documentation page for the internals.