More predictible parsing with escape_raw_string
Due to the initial parsing of the string before the macro sees it,
followed by the use of `unescape_string`, you currently need to double
up trailing backslashes. We can fix this by calling escape_raw_string in
advance.

This change may cause undesirable side-effects with less comply used
macro forms (e.g. `@styled_str "\\"`), but we'll deal with that if and
when it comes up.