Adjust to changed annotations function return form
For consistency, calling `annotations` on an AnnotatedString, whether an
index or range is specified, now always returns a vector of
region-annotation tuples.

Previously, when an index was specified, only the annotations would be
returned. Thus, code that relies on this previous behaviour must now
apply `last` to each element of the returned vector.