Introduce specialised write for AnnotatedIOBuffer
In https://github.com/JuliaLang/julia/commit/f117a500ca93, the
AnnotatedIOBuffer type was introduced. If you write from an
AnnotatedIOBuffer to another kind of IO, currently all styling
information is lost.

This seems undesirable, and we can make it work more like writing an
AnnotatedString to IO ... by converting this to an AnnotatedString write
within a specialised write method.

The conversion surely adds overhead, but it's a very staightforward way
to make this work. If this ends up being a legitimate performance
concern, we can always revisit this and either:
- Add a specialised code path for AnnotatedIOBuffer
- Try to make the current AnnotatedString-oriented methods more
  generic (somehow)