Refactor stylemacro into a submodule
Instead of having a monolithic ~600 line macro, we create a submodule
"StyledMarkup" and extract all of the inner functions of the macro into
top-level function of the submodule, creating a State type along the
way.

In my eyes, this is equivalent in terms of code organisation, but should
allow for a "functional" form of the styled"" macro with minimal
changes.