Implement new functional styled markup interpreter
Now that we have reorganised the styled"" macro, we can take a shot at
implementing a functional variant of the macro.

This allows for annotation markup in plain strings (e.g. sourced from a
user), to be converted into styled AnnotatedStrings without the use of
`eval`, opening up new use cases.