Introduce isnextchar convenience func to parser
Checking the value of the next char is a very frequent operation, and
one must always be careful to confirm that the stateful char iterator is
not empty before doing so. To improve readability, clarity, and make it
harder to forget this imported check, I'm introducing a helper function
for this task.