Use the redmean colour distance in 8-bit approx
When seeing if the nearest grey is closer to the target colour than the
nearest 6-cube colour, the reference Tmux implementation used the sRGB
l2 distance, which we know to be an awful metric for colour. It's only a
little more effort to use the "redmean" sRGB distance function (see
https://www.compuphase.com/cmetric.htm), which effectively acts as a
cheap approximation of the l2 dinstance in LAB space.