Overly aggressive color names check in styled strs
Thanks to the particular colour model used by faces, while only a
particular list of colours has default ANSI and HTML values, it is
entirely extensible.

Thus, the colour name check performed when validating styled markup
strings is actually overly-aggressive. If we work out a way to do
warnings, it might be worth developing another check for this, but for
now I think it's better not to gaslight users.