Mistyped font attribute as "face" in the parser
This is particularly annnoying since it's a essentially a minor typo
that made it impossible to specify the font as an inline face
specification. It's a little hard to check the providence of this line,
but I'm guessing it would form part of an argument in favour of getting
more sleep instead of staying up late trying to finish things.