Allow color strings as in Face constructor
Given the forms accepted by the "inline face" attribute in the styled
markup:

    styled"{(fg=blue):text}"
    styled"{(fg=#4063d8):text}"

It seems natural to some people (who have been confused by this and
asked me about it) to assume that the noted support for interpolated
values in inline face attributes means that the strings "blue" and
"#4063d8" would work equivalently.

However, this is not the case, since interpolated values are passed
directly to the Face(...) constructor, and that does not support string
colour values.

As such, the way to make using a string representation work via
interpolation is to do:

    styled"{(fg=$(parse(StyledStrings.SimpleColor, \"blue\")):text}"
    styled"{(fg=$(parse(StyledStrings.SimpleColor, \"#4063d8\"))):text}"

It's fair to say that this seems a tad over-complicated, and upon
further reflection I think it makes sense to support this use case,
given how important I think having styled markup behave intuitively is.

The most obvious way to enable this use-case is to simply add another
"colour interpretation" case to the Face(...) constructor, and so I've
done exactly that.