Try fixing the non-stdlib tests via refactor
In https://github.com/JuliaLang/julia/pull/53922 it was revealed that
simply gating the supposition-requiring test behind an if statement was
insufficient when a macro is involved in the branch not taken.

To try to make this approach work, the current conditional
fuzzer-running code was refactored to make the the macro and other code
requiring Supposition be nicely contained in a file that's only loaded
when NON_STDLIB_TESTS is true. Put another way, we've hoisted the
conditional running from an "is the code run?" if-branch to a
file-loading if-branch.