Improve inference with a function over a closure
The previous code pattern using `takewhile` relied on Julia's type
inference to produce correct results. In particular, it relied on
`takewhile(f, x) |> collect` to correctly infer to `Vector{Pair{<:Any,
Char}}`. However, type inference in Julia is an optimization, and should
not be relied on to produce the correct answer. Instead, use a normal
function to replace the functional style of programming.