Quick fix for 4d04102adf0d (Optimised SimpleColor)
Despite it "looking right" and a quick test in the REPL, it turns out I
put the _ in the wrong spot when rewriting the NTuple-based modification
to Kristoffer's commit.

Ooops. Anyway, that's fixed now, and it occurs to me that for
correctness we may as well call "htol" because we assume the byte order
is little-endian when using the results of reinterpret (even though it
will almost always be a no-op).