Small optimisation to single-interpolation styled
When a single styled"{$face:$content}" form is used, this optimisation
makes the macroexpansion equivalent to handwritten code for this case,
as the annotatedstring_optimize! call is superfluous when there is only
one part.

While we're at it, we may as well make annotatedstring_optimize! more
efficient for the zero and one annotation cases.