Mark API as Experimental
Given the ongoing discussion about how exactly bits of this library
should work (namely the Faces system), it's worth declaring that the API
is experimental, so it can stabilize while also being used across the
stdlib.

The mention of AnnotatedStrings is slightly premature, but we're about
to bring that in as part of what StyledStrings claims API responsibility
for. This is in many ways the main motivation for this notice.