Less conservative char escaping
With the !isempty check, terminal newlines as appear in

    styled"stuff\
    "

or similar are not eaten, and so parsed as "stuff\\\n".

This differs to the EBNF and standard string parsing (which we aim to
emulate), which parses

    "stuff\
    "

as "stuff".

While the !isempty check is a nice sign that I wa sbeing careful writing
the code here, it's actually safe when the state is empty, and no errors
appear in the syntax fuzzing tests.

So, all in all, it seems like we're better off without the !isempty check.