Correct ANSI printing of alternating bold/dim
The way that bold/dim work in TTYs is a little weird. They're two
channels of information (text can be bold and dim, just one, or either),
but also sort of one channel of information (weight). Dim can also be a
light font weight, which is how we frame/interpret it here. To avoid
unexpected behaviour, we clear the bold/dim status before re-setting it
when changing between the two.