Adapt to current_terminfo becoming OncePerProcess
Julia base is changing current_terminfo from a simple global variable to
a OncePerProcess type. This commit updates all usage sites to call
current_terminfo() as a function instead of accessing it directly.

This is the counterpart to
<https://github.com/JuliaLang/julia/pull/58854>.

For tests, we need to directly manipulate the OncePerProcess' internal
value field, which relies on undocumented internals but is necessary for
testing different terminal configurations without restarting the
process.