Align the semantics and handling of default fg/bg
The intended semantics of the default face is that it represents the
typesetting context that the AnnotatedString can expect to be placed in.
As such, when the value of a Face's fg/bg is the same as the default
value, we should take this to mean that the fg/bg can be applied by
restoring the default style (using the appropriate ANSI codes or CSS
declaration).

To model this appropriately, we can't make do with setting the default
foreground and background to ":default". Adding a :foreground and
:background face won't quite do the trick either. Instead, we dedicate
our basic color expectations in a new FACES attribute: basecolors. I
expect to make use of this new design soon to help implement colour
blending.

For now, the HTML colour codes are put there, along with two special
colors: foreground, and background. These are then used in the default
face, and let us disambiguate usage.